avm2: Store slot and disp ids as usize internally
Previously, they were being stored as u32s

+++ core/src/avm2/object/script_object.rs
@@ -286,22 +286,18 @@ impl<'gc> ScriptObjectWrapper<'gc> {
     }
 
     #[inline(always)]
-    pub fn get_slot(self, id: u32) -> Value<'gc> {
+    pub fn get_slot(self, id: usize) -> Value<'gc> {
         self.0
             .slots
-            .get(id as usize)
+            .get(id)
             .cloned()
             .map(|s| s.get())
             .expect("Slot index out of bounds")
     }
 
     /// Set a slot by its index.
-    pub fn set_slot(self, id: u32, value: Value<'gc>, mc: &Mutation<'gc>) {
-        let slot = self
-            .0
-            .slots
-            .get(id as usize)
-            .expect("Slot index out of bounds");
+    pub fn set_slot(self, id: usize, value: Value<'gc>, mc: &Mutation<'gc>) {
+        let slot = self.0.slots.get(id).expect("Slot index out of bounds");
 
         Gc::write(mc, self.0);
         // SAFETY: We just triggered a write barrier on the Gc.
@@ -310,8 +306,8 @@ impl<'gc> ScriptObjectWrapper<'gc> {
     }
 
     /// Retrieve a bound method from the method table.
-    pub fn get_bound_method(self, id: u32) -> Option<FunctionObject<'gc>> {
-        self.bound_methods().get(id as usize).and_then(|v| *v)
+    pub fn get_bound_method(self, id: usize) -> Option<FunctionObject<'gc>> {
+        self.bound_methods().get(id).and_then(|v| *v)
     }
 
     pub fn has_own_dynamic_property(self, name: &Multiname<'gc>) -> bool {
@@ -370,16 +366,16 @@ impl<'gc> ScriptObjectWrapper<'gc> {
     pub fn install_bound_method(
         self,
         mc: &Mutation<'gc>,
-        disp_id: u32,
+        disp_id: usize,
         function: FunctionObject<'gc>,
     ) {
         let mut bound_methods = self.bound_methods_mut(mc);
 
-        if bound_methods.len() <= disp_id as usize {
-            bound_methods.resize_with(disp_id as usize + 1, Default::default);
+        if bound_methods.len() <= disp_id {
+            bound_methods.resize_with(disp_id + 1, Default::default);
         }
 
-        *bound_methods.get_mut(disp_id as usize).unwrap() = Some(function);
+        *bound_methods.get_mut(disp_id).unwrap() = Some(function);
     }
 
     /// Get the `Class` for this object.
